+++ controllers/UsersController.php
@@ -2,18 +2,19 @@
 
 namespace app\controllers;
 
-use app\models\ResetPasswordForm;
+use app\components\BaseController;
 use Yii;
 use app\models\User;
 use app\models\UserSearch;
-use yii\web\Controller;
+use yii\web\ForbiddenHttpException;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
+use app\models\ResetPasswordForm;
 
 /**
  * UsersController implements the CRUD actions for User model.
  */
-class UsersController extends Controller
+class UsersController extends BaseController
 {
     /**
      * {@inheritdoc}
@@ -33,16 +34,21 @@ class UsersController extends Controller
     /**
      * Lists all User models.
      * @return mixed
+     * @throws ForbiddenHttpException
      */
     public function actionIndex()
     {
-        $searchModel = new UserSearch();
-        $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
+        if (Yii::$app->user->can('view_user')) {
+            $searchModel = new UserSearch();
+            $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
-        return $this->render('index', [
-            'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
-        ]);
+            return $this->render('index', [
+                'searchModel' => $searchModel,
+                'dataProvider' => $dataProvider,
+            ]);
+        } else {
+            throw new ForbiddenHttpException('You are not allowed to access this page');
+        }
     }
 
     /**
@@ -54,7 +60,7 @@ class UsersController extends Controller
     public function actionView($id)
     {
         return $this->render('view', [
-            'model' => $this->findModel($id),
+            'model' => $this->findModel($id)
         ]);
     }
 
@@ -62,18 +68,23 @@ class UsersController extends Controller
      * Creates a new User model.
      * If creation is successful, the browser will be redirected to the 'view' page.
      * @return mixed
+     * @throws ForbiddenHttpException
      */
     public function actionCreate()
     {
-        $model = new User();
+        if (Yii::$app->user->can('show-user')) {
+            $model = new User();
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->id]);
+            }
+
+            return $this->render('create', [
+                'model' => $model,
+            ]);
+        } else {
+            throw new ForbiddenHttpException('You are not allowed to access this page');
         }
-
-        return $this->render('create', [
-            'model' => $model,
-        ]);
     }
 
     /**
@@ -101,15 +112,19 @@ class UsersController extends Controller
      * If deletion is successful, the browser will be redirected to the 'index' page.
      * @param integer $id
      * @return mixed
+     * @throws ForbiddenHttpException
      * @throws NotFoundHttpException if the model cannot be found
      * @throws \Throwable
      * @throws \yii\db\StaleObjectException
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
-
-        return $this->redirect(['index']);
+        if (Yii::$app->user->can('show-user')) {
+            $this->findModel($id)->delete();
+            return $this->redirect(['index']);
+        } else {
+            throw new ForbiddenHttpException('You are not allowed to access this page');
+        }
     }
 
     /**
@@ -128,19 +143,25 @@ class UsersController extends Controller
         throw new NotFoundHttpException('The requested page does not exist.');
     }
 
-    public function actionResetPassword() {
+    /**
+     * @param $id
+     * @return string|\yii\web\Response
+     * @throws NotFoundHttpException
+     */
+    public function actionResetPassword($id)
+    {
         $model = new ResetPasswordForm();
+        $user = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post())) {
-            $user = $this->findModel(Yii::$app->user->id);
-
             // Here, we run our validation rules on the model
             if ($model->validate()) {
-                $user->setPassword($model->password);
+                $user->setPassword($model->newPassword);
                 $user->save();
                 return $this->redirect('index');
             }
         }
-        return $this->render('update', ['model' => $model]);
+        $model->password = $user->getPassword();
+        return $this->render('update', ['model' => $model, 'user' => $user]);
     }
 }
